+++ CHANGELOG.md
@@ -8,18 +8,24 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 ## Unreleased
 ### Added
 -
+- 
+- Masked backtest ([#613](https://github.com/tinkoff-ai/etna/pull/613))
+- 
+- 
+- 
 - Add plot_periodogram ([#606](https://github.com/tinkoff-ai/etna/pull/606))
 -
 -
 - Add prediction_actual_scatter_plot ([#610](https://github.com/tinkoff-ai/etna/pull/610))
 - Add plot_holidays ([#624](https://github.com/tinkoff-ai/etna/pull/624))
 
+
 ### Changed
 -
 -
 -
 -
--
+- Add in_column to plot_anomalies, plot_anomalies_interactive ([#618](https://github.com/tinkoff-ai/etna/pull/618))
 -
 
 ### Fixed